📦 brawlstars.js@1.0.6 update
Well, I have not tested this code, but it looks like to be a fine shorter version
The **brawlstars.js** library has been updated. It includes some fixes

I will probably, in the future, add a **userInfo.getClubRole()** function. Thanks for the *suggestion*
Have a nice day! *Don't forget to update it if you plan to use my code*